Fix an integer division issue with MagnitudeIntBuilder
Fixes #25.

Python 2.7.x and Python 3.x handle integer division differently.
The former returns the integer part if two integers are divided,
while the latter always returns a float.

+++ dragonfly/language/base/integer_internal.py
@@ -112,8 +112,8 @@ class MagnitudeIntBuilder(IntBuilderBase):
         if min >= max: return None
 
         # Calculate ranges of multipliers and remainders.
-        first_multiplier = min / self._factor
-        last_multiplier  = (max - 1) / self._factor + 1
+        first_multiplier = int(min / self._factor)
+        last_multiplier  = int((max - 1) / self._factor + 1)
         first_remainder  = min % self._factor
         last_remainder   = max % self._factor
         if last_remainder == 0:
